% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggr.R
\name{aggr_history_editor_count}
\alias{aggr_history_editor_count}
\title{Aggregate page history data to editor edit counts}
\usage{
aggr_history_editor_count(history_data)
}
\arguments{
\item{history_data}{data.frame}
}
\description{
Aggregate page history data to editor edit counts
}
\examples{
library(tropr)

.url <- "http://tvtropes.org/pmwiki/pmwiki.php/Characters/LittleWitchAcademia"
\dontrun{
hist_content <- trope_history(.url)
.summary <- aggr_history_editor_count(hist_content)
}
}
