% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tropr.R
\name{as.data.frame.tropr.content}
\alias{as.data.frame.tropr.content}
\title{Convert TV Trope content to data frame}
\usage{
\method{as.data.frame}{tropr.content}(content, category_name = "main")
}
\arguments{
\item{content}{\code{tropr.content} object}

\item{category_name}{a default name for trope category}
}
\value{
\code{data.frame} with tv trope contents
}
\description{
Convert TV Trope content to data frame
}
\examples{
\dontrun{
.url <- "http://tvtropes.org/pmwiki/pmwiki.php/Main/Hikikomori"
content <- trope_content(.url)
# Returns \\code{data.frame}
content <- as.data.frame(node)
}
}
