% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tropr.R
\name{trope_content}
\alias{trope_content}
\title{Get xml nodeset}
\usage{
trope_content(.url)
}
\arguments{
\item{.url}{TV Tropes page url}
}
\value{
\code{xml_nodeset} if it exists, it will return \code{xml_nodeset}
  which contains TV Tropes page content, otherwise it will show error msg
}
\description{
You can use \code{trope_node} with TV Tropes URL to get \code{xml_nodeset},
and \code{as.data.frame()} converts \code{xml_nodeset} to \code{data.frame}
}
\examples{
\dontrun{
.url <- "http://tvtropes.org/pmwiki/pmwiki.php/Main/SenseiChan"
# It will return xml nodeset
content <- trope_content(.url)
}
}
