% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{trope_data}
\alias{trope_data}
\title{Fetch trope data}
\usage{
trope_data(trope_urls, cache_dir = tempdir(), sleep = 0.5, verbose = F)
}
\arguments{
\item{trope_urls}{url of tv trope pages}

\item{cache_dir}{a directory for data caching}

\item{sleep}{wait time between queries}

\item{verbose}{verbosity option}
}
\value{
\code{data.frame} which contains trope data
}
\description{
Fetch trope data
}
\examples{
library(tropr)

.urls <- c("http://tvtropes.org/pmwiki/pmwiki.php/Main/SenseiChan",
           "http://tvtropes.org/pmwiki/pmwiki.php/Main/YouAreBetterThanYouThinkYouAre")
res <- trope_data(.urls)
}
