% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tropr.R
\name{trope_node}
\alias{trope_node}
\title{Get trope xml nodeset}
\usage{
trope_node(.url)
}
\arguments{
\item{.url}{tvtrope page url}
}
\value{
\code{xml_nodeset} which contains TV Tropes page content
}
\description{
Get trope xml nodeset
}
