% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{trope_redirect_to}
\alias{trope_redirect_to}
\title{Get the redirected urls of given trope urls}
\usage{
trope_redirect_to(trope_urls, cache_dir = tempdir(), sleep = 0.5,
  verbose = F)
}
\arguments{
\item{trope_urls}{url of tv trope pages}

\item{cache_dir}{a directory for data caching}

\item{sleep}{wait time between queries}

\item{verbose}{verbosity option}
}
\value{
\code{data.frame} which contains the redirected urls of trope urls
}
\description{
Get the redirected urls of given trope urls
}
\examples{
library(tropr)

.urls <- c("http://tvtropes.org/pmwiki/pmwiki.php/Main/SenseiChan",
           "http://tvtropes.org/pmwiki/pmwiki.php/Main/YouAreBetterThanYouThinkYouAre")
\dontrun{
res <- trope_redirect_to(.urls)
}
}
