% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{trope_cache}
\alias{trope_cache}
\title{Prepare cache for trope data}
\usage{
trope_cache(urls, depth = 1, trope_cache_dir = tempdir(),
  redirect_to_cache_dir = tempdir(), sleep = 0.5, verbose = T)
}
\description{
1) redirect_to handling
2) trope data fetching handling
}
