% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tropr.R
\name{trope_content}
\alias{trope_content}
\title{Get trope content data frame}
\usage{
trope_content(node)
}
\arguments{
\item{node}{\code{xml_nodeset} object}
}
\value{
\code{data.frame} with tv trope contents
}
\description{
Get trope content data frame
}
