% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tropr.R
\name{trope_history}
\alias{trope_history}
\title{Get TV Trope page history}
\usage{
trope_history(.url)
}
\arguments{
\item{.url}{TV Tropes page url}
}
\value{
\code{data.frame} time-series data with edit counters
}
\description{
Get TV Trope page history
}
\examples{
library(tropr)

.url <- "http://tvtropes.org/pmwiki/pmwiki.php/Characters/LittleWitchAcademia"
hist_content <- trope_history(.url)
}
